refactor: split evaluation.ts (876 lines) into 4 focused modules
Split domain evaluation queries into:
- evaluation/types.ts (278 lines) — types + Pothos object registrations
- evaluation/helpers.ts (295 lines) — parsing, status derivation, transforms
- evaluation/queries.ts (337 lines) — 6 GraphQL query fields
- evaluation/index.ts (3 lines) — barrel

All 7 existing tests pass. Zero behavioral changes.

Adversarial reviews:
- Codex: 5/5 PASS
- Gemini: 5/5 PASS
