+++ cloud/apps/api/src/graphql/queries/domain/index.ts
@@ -1,5 +1,5 @@
 import './types.js';
 import './catalog.js';
 import './planning.js';
-import './evaluation.js';
+import './evaluation/index.js';
 import './analysis.js';
